Aggornamento classi di testing per coverage e modifica bean e general

+++ src/main/java/myJava/model/professore/ReceivementManager.java
@@ -3,19 +3,31 @@ package myJava.model.professore;
 import myJava.control.connection.DriverManagerConnectionPool;
 import myJava.model.beans.Ricevimento;
 import myJava.model.beans.Studente;
+import org.mockito.internal.verification.Times;
+import sun.java2d.pipe.SpanShapeRenderer;
 
 import java.security.spec.ECField;
+
 import java.sql.*;
 import java.text.DateFormat;
+import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
+import java.util.zip.DataFormatException;
+
 
 public class ReceivementManager {
 
 
-    public boolean creaRicevimento(Ricevimento ricevimento)throws SQLException {
+    public boolean creaRicevimento(Ricevimento ricevimento)throws SQLException ,ParseException{
 
+
+        if(ricevimento.getIdRicevimento()==0||!checkData(ricevimento.getData())||!checkOrario(ricevimento.getOrarioFine())||!checkOrario(ricevimento.getOrarioInizio())||!checkLuogo(ricevimento.getLuogo())){
+
+            return false;
+        }
         Connection connection = null;
         PreparedStatement preparedStatement = null;
 
@@ -34,7 +46,7 @@ public class ReceivementManager {
             preparedStatement.executeUpdate();
 
             connection.commit();
-            System.out.println("ho commitatto");
+
             return true;
         } catch (Exception e)
         {
@@ -49,13 +61,18 @@ public class ReceivementManager {
 
     public boolean eliminaRicevimento(Ricevimento ricevimento)throws SQLException {
 
+if(ricevimento.getIdRicevimento()==0){
+    return false;
+}
         Connection conn;
-
         conn = DriverManagerConnectionPool.getConnection();
         try {
             PreparedStatement preparedStatement = conn.prepareStatement("delete from ricevimento where idRicevimento=?");
             preparedStatement.setInt(1, ricevimento.getIdRicevimento());
-            preparedStatement.executeUpdate();
+            if(preparedStatement.executeUpdate()==0){//non è stato cancellato alcun ricevimento
+                throw new Exception();
+
+            }
             conn.commit();
 
             return true;
@@ -81,7 +98,9 @@ try {
     preparedStatement.setString(5,ricevimento.getData());
     preparedStatement.setInt(6,ricevimento.getIdProfessore());
     preparedStatement.setInt(7, ricevimento.getIdRicevimento());
-    preparedStatement.executeUpdate();
+    if(preparedStatement.executeUpdate()==0){
+        throw new Exception();
+    }
     conn.commit();
     preparedStatement.close();
     conn.close();
@@ -94,9 +113,12 @@ try {
 }
     }
 
-    public Ricevimento visualizzaRicevimento( String orarioInizio, String orarioFine, String dataR)throws SQLException{
+    public Ricevimento visualizzaRicevimento( String orarioInizio, String orarioFine, String dataR)throws SQLException,ParseException{
 
+if(!checkOrario(orarioInizio)||!checkOrario(orarioFine)||!checkData(dataR)){
 
+    return null;
+}
         Connection conn=null;
         PreparedStatement preparedStatement =null;
 try {
@@ -129,7 +151,10 @@ try {
     }
 
     public boolean registraPresenza(int idStudente)throws SQLException{
+if(idStudente==0){
 
+    return false;
+}
         Connection conn=null;
         PreparedStatement preparedStatement=null;
 try {
@@ -137,7 +162,10 @@ try {
     preparedStatement = conn.prepareStatement("update prenotazione  set prenotazione.presenza=1 where prenotazione.idStudente=?");
 
     preparedStatement.setInt(1, idStudente);
-    preparedStatement.executeUpdate();
+   if(preparedStatement.executeUpdate()==0){
+
+       throw new Exception();
+   }
     conn.commit();
     return true;
 }catch (Exception e){
@@ -152,25 +180,43 @@ try {
     }}
 
     public boolean registraAssenza(int idStudente)throws SQLException{
+        if(idStudente==0){
 
+            return false;
+        }
         Connection conn=null;
         PreparedStatement preparedStatement=null;
+try {
+    conn = DriverManagerConnectionPool.getConnection();
 
-        conn=DriverManagerConnectionPool.getConnection();
+    preparedStatement = conn.prepareStatement("update studente set numAssenza=numAssenza+1 where idStudente=?");
+    preparedStatement.setInt(1, idStudente);
+    if (preparedStatement.executeUpdate() == 0) {
 
-        preparedStatement=conn.prepareStatement("update studente set numAssenza=numAssenza+1 where idStudente=?");
-        preparedStatement.setInt(1,idStudente);
-        preparedStatement.executeUpdate();
-        conn.commit();
-        preparedStatement.close();
-        conn.close();
-        return true;
+        throw new Exception();
+    }
+    conn.commit();
+
+    return true;
+}catch (Exception e){
+    System.err.println("Got an exception! ");
+    System.err.println(e.getMessage());
+
+    return false;
+
+}
+
 
 
 
     }
     public List<Studente> visualizzaStudenti(Ricevimento ricevimento) throws SQLException{
+if(ricevimento == null || ricevimento.getIdRicevimento()==0){
 
+
+    return null;
+
+}
         Connection connection = null;
 
         List<Studente> students =new ArrayList<>();
@@ -202,9 +248,12 @@ try {
 
     }
 
-
     //get Ricevimento byId
     public Ricevimento getRicevimentoById(int idRicevimento){
+        if(idRicevimento==0){
+
+            return null;
+        }
         Connection connection = null;
         Ricevimento ricevimento = new Ricevimento();
         DateFormat simpleDateFormat = new SimpleDateFormat("HH:mm");
@@ -231,6 +280,45 @@ try {
             e.printStackTrace();
         }
         return ricevimento;
+    }
+
+    private boolean checkData(String data){
+        if(data.matches("\\d{4}-\\d{2}-\\d{2}")){
+
+            return true;
+        }
+        else return false;
+    }
+
+    private boolean checkOrario(String orario)throws ParseException {
+        if(orario.matches("\\d{2}:\\d{2}:\\d{2}")){
+                String dataFittizzia="2019-12-12 ";
+       String dataCompleta= dataFittizzia + orario;
+SimpleDateFormat format=new SimpleDateFormat("yyyy-MM-dd hh:mm:ss");
+ Date parsedDate=format.parse(dataCompleta);
+ Timestamp timeStamp=new java.sql.Timestamp( parsedDate.getTime());
+ parsedDate=format.parse(dataFittizzia+"20:00:00");
+ Timestamp orarioLimiteInferiore=new java.sql.Timestamp( parsedDate.getTime());
+ parsedDate=format.parse(dataFittizzia+"09:00:00");
+ Timestamp orarioLimiteSuperiore=new Timestamp(parsedDate.getTime());
+
+ if(timeStamp.before(orarioLimiteInferiore)&&timeStamp.after(orarioLimiteSuperiore)){
+            return true;
+
+        }
+ return false;
+        }
+else
+    return false;
+    }
+
+    private boolean checkLuogo(String luogo){
+
+        if(!luogo.equals("")){
+
+            return true;
+        }
+        return false;
     }
 
 }
